sdcard->bfin boot ok

+++ common/fat_io_lib/fat_types.h
@@ -7,7 +7,7 @@
 #ifndef FATFS_NO_DEF_TYPES
 typedef unsigned char uint8;
 typedef unsigned short uint16;
-typedef unsigned int uint32;
+typedef unsigned long int uint32;
 #endif
 
 #ifndef NULL
